fix: 뉴스레터 파이프라인 상태 트랜잭션 분리
